Fix issue with GithubContents
Apparently, Github stopped returning content in (some?) [API responses](https://api.github.com/repos/codeforkyana/power-outage-data/contents/lgeku/outages.json), which means GithubContents started failing to download `outage.json`. It probably has to do with file size. This fix looks for a `download_url` in the API response and uses it if it's present and `content` is empty or missing.